Removed PyTorch dependencies
