+++ luxonis_ml/embeddings/__init__.py
@@ -16,18 +16,12 @@ with guard_missing_extra("embedd"):
         QdrantManager,
         WeaviateAPI,
         VectorDBAPI,
-        export_model_onnx,
         extend_output_onnx,
         extend_output_onnx_overwrite,
-        extract_embeddings,
-        extract_embeddings_onnx,
-        extract_embeddings_onnx_LFS,
-        generate_embeddings,
-        load_embeddings,
-        load_model,
         load_model_onnx,
-        load_model_resnet50_minuslastlayer,
-        save_embeddings,
+        save_model_onnx,
+        extract_embeddings,
+        generate_embeddings,
     )
 
 __all__ = [
@@ -39,10 +33,8 @@ __all__ = [
     "calculate_similarity_matrix",
     "find_representative_greedy",
     "find_representative_kmedoids",
-    "load_model_resnet50_minuslastlayer",
-    "load_model",
-    "export_model_onnx",
     "load_model_onnx",
+    "save_model_onnx",
     "extend_output_onnx",
     "extend_output_onnx_overwrite",
     "QdrantManager",
@@ -50,9 +42,5 @@ __all__ = [
     "WeaviateAPI",
     "VectorDBAPI",
     "extract_embeddings",
-    "extract_embeddings_onnx",
-    "extract_embeddings_onnx_LFS",
-    "save_embeddings",
-    "load_embeddings",
     "generate_embeddings",
 ]
