+++ luxonis_ml/embeddings/utils/__init__.py
@@ -1,28 +1,18 @@
-from .embedding import (
-    extract_embeddings,
-    extract_embeddings_onnx,
-    extract_embeddings_onnx_LFS,
-    load_embeddings,
-    save_embeddings,
-)
+from .embedding import extract_embeddings
 from .ldf import generate_embeddings
 from .model import (
-    export_model_onnx,
     extend_output_onnx,
     extend_output_onnx_overwrite,
-    load_model,
     load_model_onnx,
-    load_model_resnet50_minuslastlayer,
+    save_model_onnx,
 )
 from .qdrant import QdrantAPI, QdrantManager
 from .weaviate import WeaviateAPI
 from .vectordb import VectorDBAPI
 
 __all__ = [
-    "load_model_resnet50_minuslastlayer",
-    "load_model",
-    "export_model_onnx",
     "load_model_onnx",
+    "save_model_onnx",
     "extend_output_onnx",
     "extend_output_onnx_overwrite",
     "QdrantManager",
@@ -30,9 +20,5 @@ __all__ = [
     "WeaviateAPI",
     "VectorDBAPI",
     "extract_embeddings",
-    "extract_embeddings_onnx",
-    "extract_embeddings_onnx_LFS",
-    "save_embeddings",
-    "load_embeddings",
     "generate_embeddings"
 ]
