+++ luxonis_ml/embeddings/utils/embedding.py
@@ -42,109 +42,78 @@ Dependencies:
     - onnxruntime
     - onnx
 """
-from PIL import Image
-from io import BytesIO
-from typing import Tuple, List
-
+import cv2
+import numpy as np
+from typing import List, Tuple, Callable
 import onnxruntime as ort
-import torch
-import torchvision.transforms as transforms
+from io import BytesIO
 
-from luxonis_ml.data import LuxonisDataset
 from luxonis_ml.utils import LuxonisFileSystem
 
-def extract_embeddings(
-    model: torch.nn.Module, data_loader: torch.utils.data.DataLoader
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """Extract embeddings from the given PyTorch model."""
-    embeddings = []
-    labels = []
 
-    with torch.no_grad():
-        for images, batch_labels in data_loader:
-            outputs = model(images)
-            embeddings.extend(outputs.squeeze())
-            labels.extend(batch_labels)
-
-    return torch.stack(embeddings), torch.tensor(labels)
-
-
-def extract_embeddings_onnx(
-    ort_session: ort.InferenceSession,
-    data_loader: torch.utils.data.DataLoader,
-    output_layer_name: str,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """Extract embeddings from the given ONNX model."""
-    embeddings = []
-    labels = []
-
-    with torch.no_grad():
-        for images, batch_labels in data_loader:
-            ort_inputs = {ort_session.get_inputs()[0].name: images.numpy()}
-            outputs = ort_session.run([output_layer_name], ort_inputs)[0]
-            embeddings.extend(torch.from_numpy(outputs).squeeze())
-            labels.extend(batch_labels)
-
-    return torch.stack(embeddings), torch.tensor(labels)
+def preprocess_image_cv2(img: np.ndarray) -> np.ndarray:
+    """Custom preprocessing function provided by the user."""
+    # Resize and convert BGR to RGB
+    img = cv2.resize(img, (224, 224))
+    img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+    
+    # Normalize the image
+    img = img.astype(np.float32) / 255.0
+    mean = np.array([0.485, 0.456, 0.406])
+    std = np.array([0.229, 0.224, 0.225])
+    img_normalized = (img - mean) / std
+    
+    # Change to channel-first format
+    img_transposed = img_normalized.transpose(2, 0, 1)
+    
+    return img_transposed
 
 def get_image_tensors_from_LFS(
     image_paths: List[str],
-    transform: transforms.Compose,
+    preprocess_function: Callable[[np.ndarray], np.ndarray],
     lfs: LuxonisFileSystem,
-) -> Tuple[torch.Tensor, List[int]]:
+) -> Tuple[np.ndarray, List[int]]:
     tensors = []
     successful_ixs = []
 
     for i, path in enumerate(image_paths):
         try:
+            # Read image data into a byte buffer
             buffer = lfs.read_to_byte_buffer(remote_path=path).getvalue()
-            image = Image.open(BytesIO(buffer)).convert('RGB')
-            tensor = transform(image)
-            tensors.append(tensor)
+            image = np.frombuffer(buffer, dtype=np.uint8)
+            image = cv2.imdecode(image, cv2.IMREAD_COLOR)
+            processed_image = preprocess_function(image)
+            tensors.append(processed_image)
             successful_ixs.append(i)
-        except:
-            print("Error occured while processing image: ", path)
+        except Exception as e:
+            print(f"Error occurred while processing image: {path}, Error: {e}")
             continue
+    
+    return np.stack(tensors), successful_ixs
 
-    return torch.stack(tensors), successful_ixs
-
-def extract_embeddings_onnx_LFS(
+def extract_embeddings(
     image_paths: List[str],
     ort_session: ort.InferenceSession,
-    transform: transforms.Compose,
-    lfs: LuxonisFileSystem,
+    lfs: LuxonisFileSystem, 
+    preprocess_function: Callable[[np.ndarray], np.ndarray],
     output_layer_name: str = "/Flatten_output_0",
     batch_size: int = 64,
-) -> Tuple[torch.Tensor, List[int]]:
+) -> Tuple[List[List[float]], List[int]]:
     embeddings = []
     successful_ixs = []
 
+    if preprocess_function is None:
+        preprocess_function = preprocess_image_cv2
+
     for i in range(0, len(image_paths), batch_size):
         batch_paths = image_paths[i:i + batch_size]
-        batch_tensors, batch_ixs = get_image_tensors_from_LFS(batch_paths, transform, lfs)
+        batch_tensors, batch_ixs = get_image_tensors_from_LFS(batch_paths, preprocess_function, lfs)
 
-        # Extract embeddings using ONNX
-        ort_inputs = {ort_session.get_inputs()[0].name: batch_tensors.numpy()}
+        # Prepare inputs for ONNX runtime
+        ort_inputs = {ort_session.get_inputs()[0].name: batch_tensors.astype(np.float32)}
         ort_outputs = ort_session.run([output_layer_name], ort_inputs)[0]
 
-        embeddings.extend(torch.from_numpy(ort_outputs).squeeze())
+        embeddings.extend(ort_outputs.tolist())
+        successful_ixs.extend([i + ix for ix in batch_ixs])
 
-        batch_ixs = [i + ix for ix in batch_ixs]
-        successful_ixs.extend(batch_ixs)
-
-    return torch.stack(embeddings), successful_ixs
-
-def save_embeddings(
-    embeddings: torch.Tensor, labels: torch.Tensor, save_path: str = "./"
-):
-    """Save embeddings and labels tensors to the specified path."""
-    torch.save(embeddings, save_path + "embeddings.pth")
-    torch.save(labels, save_path + "labels.pth")
-
-
-def load_embeddings(save_path: str = "./") -> Tuple[torch.Tensor, torch.Tensor]:
-    """Load embeddings and labels tensors from the specified path."""
-    embeddings = torch.load(save_path + "embeddings.pth")
-    labels = torch.load(save_path + "labels.pth")
-
-    return embeddings, labels
+    return embeddings, successful_ixs
